predict and train.py with db quesries
