+++ lights_temp_automation/predict.py
@@ -1,4 +1,3 @@
-# lights_temp_automation/predict.py
 import numpy as np
 import pandas as pd
 import tensorflow as tf
@@ -6,6 +5,11 @@ from tensorflow import keras
 import os
 from dotenv import load_dotenv
 import datetime
+import sys
+
+# Add database directory to path to import database module
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+from database.database import db_get_sensor_data_for_prediction, db_get_light_and_temp_sensors, db_save_predicted_values
 
 
 def load_environment():
@@ -19,12 +23,10 @@ def load_environment():
         # Default to the directory one level up from this script
         project_path = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
-    # Construct paths dynamically based on the project path
+    # Construct model path dynamically based on the project path
     model_h5_path = os.path.join(project_path, 'model.h5')
-    test_csv_path = os.path.join(project_path, 'test.csv')
-    predictions_csv_path = os.path.join(project_path, 'predictions.csv')
 
-    return project_path, model_h5_path, test_csv_path, predictions_csv_path
+    return project_path, model_h5_path
 
 
 def load_model(model_h5_path):
@@ -38,36 +40,47 @@ def load_model(model_h5_path):
         exit(1)
 
 
-def load_and_preprocess_data(test_csv_path):
-    # Load test data from CSV
+def load_and_preprocess_data():
+    # Load test data from database
     try:
-        df = pd.read_csv(test_csv_path)
-        print(f"Loaded test data with {len(df)} records")
+        # Get sensor data for the last day
+        df = db_get_sensor_data_for_prediction(days=1)
+        print(f"Loaded sensor data with {len(df)} records")
+
+        # Get light and temperature sensor names
+        light_sensors, temp_sensors = db_get_light_and_temp_sensors()
 
         if len(df) < 24:
-            print("Warning: Test data contains fewer than 24 records. The model expects 24 time steps.")
+            print("Warning: Sensor data contains fewer than 24 records. The model expects 24 time steps.")
             print(f"Current record count: {len(df)}")
 
         # Keep timestamp for reference but don't use it for prediction
         timestamps = df['timestamp'].tolist()
+        last_timestamp = timestamps[-1] if timestamps else datetime.datetime.now()
 
-        # Normalize time features exactly as done in train.py
+        # Normalize time features
         df['hour'] = df['hour'] / 23.0  # Normalize hour (0-23 → 0-1)
         df['day_of_week'] = df['day_of_week'] / 6.0  # Normalize day (0-6 → 0-1)
 
-        # Normalize temperature values just like in train.py
-        df['t1'] = (df['t1'] - 20) / 10.0
-        df['t2'] = (df['t2'] - 20) / 10.0
-        df['t3'] = (df['t3'] - 20) / 10.0
-        df['t4'] = (df['t4'] - 20) / 10.0
+        # Normalize temperature values
+        for temp_sensor in temp_sensors:
+            if temp_sensor in df.columns:
+                # Use dynamic normalization based on the temperature range
+                # or use fixed normalization like in the original code
+                df[temp_sensor] = (df[temp_sensor] - 20) / 10.0
 
-        # Select the same features used in training
-        features = ['l1', 'l2', 'l3', 'l4', 'l5', 'l6', 'l7', 'l8',
-                    't1', 't2', 't3', 't4', 'hour', 'day_of_week']
+        # Select features for prediction
+        features = light_sensors + temp_sensors + ['hour', 'day_of_week']
+
+        # Make sure all required features exist
+        for feature in features:
+            if feature not in df.columns:
+                print(f"Warning: Feature {feature} not found in data. Adding with zeros.")
+                df[feature] = 0.0
+
         X = df[features].values.astype(np.float32)
 
         # Reshape input for LSTM: (1, SEQ_LEN, features)
-        # SEQ_LEN = 24 as defined in train.py
         SEQ_LEN = 24
 
         # Ensure we have exactly 24 timesteps
@@ -82,7 +95,7 @@ def load_and_preprocess_data(test_csv_path):
         # Reshape for LSTM input: (samples, timesteps, features)
         X = X.reshape(1, SEQ_LEN, len(features))
 
-        return X, timestamps[-1] if timestamps else datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+        return X, last_timestamp, light_sensors, temp_sensors
 
     except Exception as e:
         print(f"Error preprocessing data: {e}")
@@ -100,71 +113,44 @@ def make_predictions(model, X):
         exit(1)
 
 
-# def process_predictions(predictions):
-#     # Process the raw predictions to get usable values
-#
-#     # Extract light and temperature predictions
-#     # The model outputs 12 values: 8 for lights and 4 for temperature
-#     light_predictions = predictions[0, :8]  # First 8 values are for lights
-#     temp_predictions = predictions[0, 8:12]  # Last 4 values are for temperature
-#
-#     # Process light predictions - convert to integers 0-3
-#     # Map to the nearest valid level (0, 1, 2, 3)
-#     processed_lights = np.round(np.clip(light_predictions, 0, 3)).astype(int)
-#
-#     # Process temperature predictions - denormalize as per train.py
-#     # In train.py: df['t1'] = (df['t1'] - 20) / 10.0
-#     # So reverse: temp = (normalized_temp * 10.0) + 20
-#     processed_temps = (temp_predictions * 10.0) + 20.0
-#
-#     # Combine the processed predictions
-#     final_predictions = np.concatenate([processed_lights, processed_temps])
-#
-#     # Format the results
-#     results = {
-#         'lights': {f'l{i + 1}': int(processed_lights[i]) for i in range(8)},
-#         'temperatures': {f't{i + 1}': round(float(processed_temps[i]), 2) for i in range(4)}
-#     }
-#
-#     return final_predictions, results
-
-
-def process_predictions(predictions):
+def process_predictions(predictions, light_sensors, temp_sensors):
     # Process the raw predictions to get usable values
 
     # Extract light and temperature predictions
-    light_predictions = predictions[0, :8]  # First 8 values are for lights
-    temp_predictions = predictions[0, 8:12]  # Last 4 values are for temperature
+    num_lights = len(light_sensors)
+    num_temps = len(temp_sensors)
+
+    light_predictions = predictions[0, :num_lights]  # First values are for lights
+    temp_predictions = predictions[0, num_lights:num_lights + num_temps]  # Last values are for temperature
 
     # Process light predictions - convert to integers 0-3
     processed_lights = np.round(np.clip(light_predictions, 0, 3)).astype(int)
 
     # Add random variation to temperature predictions
-    # The random factor will change with each run
     random_variation = np.random.uniform(-2.0, 2.0, size=len(temp_predictions))
-
-    # Apply the random variation before denormalizing
     varied_temp_predictions = temp_predictions + random_variation * 0.2  # 0.2 controls magnitude of variation
 
-    # Process temperature predictions - denormalize as per train.py
+    # Process temperature predictions - denormalize
     processed_temps = (varied_temp_predictions * 10.0) + 20.0
 
-    # Combine the processed predictions
-    final_predictions = np.concatenate([processed_lights, processed_temps])
-
-    # Format the results
+    # Format the results as dictionaries with sensor names
     results = {
-        'lights': {f'l{i + 1}': int(processed_lights[i]) for i in range(8)},
-        'temperatures': {f't{i + 1}': round(float(processed_temps[i]), 2) for i in range(4)}
+        'lights': {light_sensors[i]: int(processed_lights[i]) for i in range(num_lights)},
+        'temperatures': {temp_sensors[i]: round(float(processed_temps[i]), 2) for i in range(num_temps)}
     }
+
+    # Combine the processed predictions (for CSV output if needed)
+    final_predictions = np.concatenate([processed_lights, processed_temps])
 
     return final_predictions, results
 
-def save_predictions(final_predictions, predictions_csv_path):
+
+def save_predictions_to_csv(final_predictions, light_sensors, temp_sensors, project_path):
     # Save predictions to CSV file
-    predictions_df = pd.DataFrame([final_predictions],
-                                  columns=["l1", "l2", "l3", "l4", "l5", "l6", "l7", "l8",
-                                           "t1", "t2", "t3", "t4"])
+    predictions_csv_path = os.path.join(project_path, 'predictions.csv')
+
+    column_names = light_sensors + temp_sensors
+    predictions_df = pd.DataFrame([final_predictions], columns=column_names)
 
     predictions_df.to_csv(predictions_csv_path, index=False, float_format='%.2f')
     print(f"Predictions saved to {predictions_csv_path}")
@@ -174,14 +160,14 @@ def main():
     print("Starting prediction process...")
 
     # Load environment and paths
-    project_path, model_h5_path, test_csv_path, predictions_csv_path = load_environment()
+    project_path, model_h5_path = load_environment()
     print(f"Project path: {project_path}")
 
     # Load the trained model
     model = load_model(model_h5_path)
 
-    # Load and preprocess test data
-    X, last_timestamp = load_and_preprocess_data(test_csv_path)
+    # Load and preprocess sensor data from database
+    X, last_timestamp, light_sensors, temp_sensors = load_and_preprocess_data()
     print(f"Preprocessed data shape: {X.shape}")
     print(f"Last timestamp: {last_timestamp}")
 
@@ -189,9 +175,14 @@ def main():
     raw_predictions = make_predictions(model, X)
 
     # Process the predictions
-    final_predictions, results = process_predictions(raw_predictions)
+    final_predictions, results = process_predictions(raw_predictions, light_sensors, temp_sensors)
+
     # Save predictions to CSV file
-    save_predictions(final_predictions, predictions_csv_path)
+    save_predictions_to_csv(final_predictions, light_sensors, temp_sensors, project_path)
+
+    # Save predictions to database
+    db_save_predicted_values(results)
+
     # Print detailed results
     print("\nPrediction results:")
     print("Light levels (0-3): ", end="")
