--- lights_temp_automation/test_db.py
@@ -1,3 +1,0 @@
-from database.database import *
-
-db_assign_module()
